CO: Move configuration interface
Conflicts:
	classes/tax/AverageTaxOfProductsTaxCalculator.php
	src/Core/ContainerBuilder.php
	src/Core/Foundation/Database/EntityManager.php
	tests/Unit/classes/tax/AverageTaxOfProductsTaxCalculatorTest.php
