+++ tests/Unit/classes/CartTest.php
@@ -26,11 +26,7 @@
 
 namespace PrestaShop\PrestaShop\Tests\Unit\Classes;
 
-use Exception;
 use PrestaShop\PrestaShop\Tests\TestCase\UnitTestCase;
-use PrestaShop\PrestaShop\Adapter\Product\PriceCalculator;
-use PrestaShop\PrestaShop\Adapter\AddressFactory;
-use PrestaShop\PrestaShop\Core\ConfigurationInterface;
 use Address;
 use Cart;
 use Order;
